remove gActiveBattler

+++ include/battle_util2.h
@@ -5,6 +5,6 @@ void AllocateBattleResources(void);
 void FreeBattleResources(void);
 void AdjustFriendshipOnBattleFaint(u8 battler);
 void SwitchPartyOrderInGameMulti(u8 battler, u8 arg1);
-u32 BattlePalace_TryEscapeStatus(u8 battler);
+u32 BattlePalace_TryEscapeStatus(u32 battler);
 
 #endif // GUARD_BATTLE_UTIL_H
